Project-v2/requirements.txt
streamlit==1.34.0
PyPDF2==3.0.1
python-docx==1.1.2
langchain-core==0.1.45
langchain-openai==0.1.3
python-docx==1.1.2